+++ src/Preprocessing/Imputer.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Phpml\Preprocessing;
 
+use Phpml\Exception\InvalidOperationException;
 use Phpml\Preprocessing\Imputer\Strategy;
 
 class Imputer implements Preprocessor
@@ -50,6 +51,10 @@ class Imputer implements Preprocessor
 
     public function transform(array &$samples): void
     {
+        if ($this->samples === []) {
+            throw new InvalidOperationException('Missing training samples for Imputer.');
+        }
+
         foreach ($samples as &$sample) {
             $this->preprocessSample($sample);
         }